pyproject.toml
[project]
name = "dashboard-v4"
version = "4.2.1"
description = "Python eInk dashboard for Raspberry Pi — calendar, weather, birthdays, and daily quotes"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "Pillow>=10.0",
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "requests>=2.28",
    "PyYAML>=6.0",
    "icalendar>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.0",
]
pi = [
    "gpiozero",
    "lgpio",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = [
    "E501",  # line length handled by formatter
    "UP007", # X | Y syntax requires 3.10+ for runtime (we use from __future__ for annotations)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = "src.data.models"
disallow_untyped_defs = true
warn_return_any = true

# Render components call Optional[FontCallable] fields (e.g. style.font_bold(size))
# that are guaranteed non-None by ThemeStyle.__post_init__, but mypy can't prove this.
# ignore_errors suppresses the pervasive "None not callable" [misc] false positives.
# Runtime correctness is verified by the smoke tests in the test suite.
[[tool.mypy.overrides]]
module = [
    "src.render.components.air_quality_panel",
    "src.render.components.birthday_bar",
    "src.render.components.diags_panel",
    "src.render.components.fuzzyclock_panel",
    "src.render.components.header",
    "src.render.components.info_panel",
    "src.render.components.message_panel",
    "src.render.components.moonphase_panel",
    "src.render.components.qotd_panel",
    "src.render.components.scorecard_panel",
    "src.render.components.sunrise_panel",
    "src.render.components.tides_panel",
    "src.render.components.timeline_panel",
    "src.render.components.today_view",
    "src.render.components.weather_full",
    "src.render.components.weather_panel",
    "src.render.components.week_view",
    "src.render.components.year_pulse_panel",
]
ignore_errors = true
